exchange_platform/requirements.txt
asgiref==3.9.0
Django==4.2.23
django-appconf==1.1.0
django-filter==25.1
django-imagekit==5.0.0
djangorestframework==3.16.0
dotenv==0.9.9
gunicorn==23.0.0
packaging==25.0
pilkit==3.0
pillow==11.3.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
